+++ src/AutoRest.CSharp/Output/Models/Types/OutputLibrary.cs
@@ -35,11 +35,6 @@ namespace AutoRest.CSharp.Output.Models.Types
         private Dictionary<Operation, ResponseHeaderGroupType>? _headerModels;
         private Dictionary<string, List<OperationGroup>> _operationGroups;
         private IEnumerable<Schema> _allSchemas;
-        public static IList<System.Type> ReferenceClassCollection = GetReferenceClassCollection();
-
-        private Dictionary<Schema, TypeProvider> SchemaMap => _models ??= BuildModels();
-
-        public Dictionary<Schema, TypeProvider> ResourceSchemaMap => _resourceModels ??= BuildResourceModels();
 
         public OutputLibrary(CodeModel codeModel, BuildContext context)
         {
@@ -57,13 +52,9 @@ namespace AutoRest.CSharp.Output.Models.Types
             }
         }
 
-        public void RebuildModelInheritance()
-        {
-            var typeOverrideMap = new Dictionary<CSharpType, CSharpType>();
+        public Dictionary<Schema, TypeProvider> SchemaMap => _models ??= BuildModels();
 
-            RebuildExactModelInheritance(SchemaMap);
-            RebuildExactModelInheritance(ResourceSchemaMap);
-        }
+        public Dictionary<Schema, TypeProvider> ResourceSchemaMap => _resourceModels ??= BuildResourceModels();
 
         public IEnumerable<TypeProvider> Models => SchemaMap.Values;
 
@@ -82,99 +73,6 @@ namespace AutoRest.CSharp.Output.Models.Types
         public IEnumerable<LongRunningOperation> LongRunningOperations => EnsureLongRunningOperations().Values;
 
         public IEnumerable<ResponseHeaderGroupType> HeaderModels => (_headerModels ??= EnsureHeaderModels()).Values;
-
-        private static IList<System.Type> GetReferenceClassCollection()
-        {
-            var assembly = Assembly.GetAssembly(typeof(AzureResourceManagerClient));
-            if (assembly is null)
-            {
-                return new List<System.Type>();
-            }
-            return assembly.GetTypes().Where(t => t.GetCustomAttributes(false).Where(a => a.GetType() == typeof(ReferenceTypeAttribute)).Count() > 0).ToList();
-        }
-
-        private void RebuildExactModelInheritance(Dictionary<Schema, TypeProvider> map)
-        {
-            foreach (var kval in map)
-            {
-                if (kval.Key is ObjectSchema objectSchema)
-                {
-                    var childObjectType = (ObjectType)kval.Value;
-                    var typeToReplace = childObjectType?.Inherits?.Implementation as ObjectType;
-                    if (typeToReplace is null)
-                    {
-                        continue;
-                    }
-
-                    var parent = GetExactMatch(typeToReplace.Type);
-                    if (parent != null)
-                    {
-                        childObjectType?.OverrideInherits(parent);
-                    }
-                }
-            }
-        }
-
-        private CSharpType? GetExactMatch(CSharpType childType)
-        {
-            foreach (System.Type parentType in ReferenceClassCollection)
-            {
-                if (IsEqual(childType, parentType))
-                {
-                    return parentType;
-                }
-            }
-            return null;
-        }
-
-        private bool IsEqual(CSharpType childType, System.Type parentType)
-        {
-            var childProperties = ((ObjectType)(childType.Implementation)).Properties.ToList();
-            List<PropertyInfo> parentProperties = parentType.GetProperties(BindingFlags.Public | BindingFlags.Instance).ToList();
-
-            if (parentProperties.Count > childProperties.Count)
-            {
-                return false;
-            }
-
-            Dictionary<string, PropertyInfo> parentDict = new Dictionary<string, PropertyInfo>();
-            foreach (var parentProperty in parentProperties)
-            {
-                parentDict.Add(parentProperty.Name, parentProperty);
-            }
-
-            foreach (var childProperty in childProperties)
-            {
-                PropertyInfo? parentProperty;
-                CSharpType childPropertyType = childProperty.Declaration.Type;
-                if (parentDict.TryGetValue(childProperty.Declaration.Name, out parentProperty))
-                {
-                    if (parentProperty.PropertyType.IsGenericType)
-                    {
-                        if (!childPropertyType.Equals(new CSharpType(parentProperty.PropertyType)))
-                            return false;
-                    }
-                    else if (parentProperty.PropertyType.FullName != $"{childPropertyType.Namespace}.{childPropertyType.Name}" &&
-                        !IsAssignable(parentProperty.PropertyType, childPropertyType))
-                    {
-                        //TODO(ADO item 5712): deal with protected setter
-                        return false;
-                    }
-                }
-                else
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        private bool IsAssignable(System.Type parentPropertyType, CSharpType childPropertyType)
-        {
-            return parentPropertyType.GetMethods().Where(m => m.Name == "op_Implicit" &&
-                m.ReturnType.FullName == $"{childPropertyType.Namespace}.{childPropertyType.Name}" &&
-                m.GetParameters().First().ParameterType == parentPropertyType).Count() > 0;
-        }
 
         private Dictionary<Operation, ResponseHeaderGroupType> EnsureHeaderModels()
         {
